+++ base/settings/tests/testthat/test.loadPath_sitePFT.R
@@ -1,12 +1,10 @@
 test_that("`loadPath.sitePFT` gives no return value for file extensions other than csv and txt", {
-  settings <- list(host = "pecan")
   path <- "base/settings.R"
-  expect_silent(loadPath.sitePFT(settings, path))
+  expect_silent(loadPath.sitePFT(path))
 })
 
 test_that("`loadPath.sitePFT` gives an error for file with number of columns not equal to 2", {
   withr::with_tempfile("tf", fileext = ".csv", {
-    settings <- list(host = "pecan")
     df <- data.frame(
       h1 = c("1", "2", "3"),
       h2 = c("a", "b", "c"),
@@ -14,7 +12,7 @@ test_that("`loadPath.sitePFT` gives an error for file with number of columns not
     )
     write.csv(df, tf, row.names = FALSE)
     expect_error(
-      loadPath.sitePFT(settings, tf),
+      loadPath.sitePFT(tf),
       "file does not have two columns."
     )
   })
@@ -22,7 +20,6 @@ test_that("`loadPath.sitePFT` gives an error for file with number of columns not
 
 test_that("`loadPath.sitePFT` works for correct format of input file",{
   withr::with_tempfile("tf", fileext = ".csv", {
-    settings <- list(host = "pecan")
     df <- data.frame(
       h1 = c("1", "2", "3"),
       h2 = c("a", "b", "c")
@@ -30,6 +27,6 @@ test_that("`loadPath.sitePFT` works for correct format of input file",{
 
     write.csv(df, tf, row.names = FALSE)
     links <- utils::read.table(tf, header = TRUE, sep = ",")
-    expect_equal(loadPath.sitePFT(settings, tf), `colnames<-`(links, c("site", "pft")))
+    expect_equal(loadPath.sitePFT(tf), `colnames<-`(links, c("site", "pft")))
   })
 })