+++ models/sipnet/tests/testthat/test.met2model.R
@@ -49,3 +49,39 @@ test_that("Missing data throws an error", {
     expect_length(list.files(".", "*.clim"), 0)
   })
 })
+
+
+test_that("clim format switch", {
+  full_nc <- system.file("test-data", "CRUNCEP.2000.nc", package = "PEcAn.utils")
+
+  outdir1 <- withr::local_tempdir()
+  outdir2 <- withr::local_tempdir()
+
+  res1 <- met2model.SIPNET(
+    in.path = dirname(full_nc),
+    in.prefix = "CRUNCEP",
+    outfolder = outdir1,
+    start_date = "2000-01-01",
+    end_date = "2000-12-31",
+    clim_format_version = "v1"
+  )
+  res2 <- met2model.SIPNET(
+    in.path = dirname(full_nc),
+    in.prefix = "CRUNCEP",
+    outfolder = outdir2,
+    start_date = "2000-01-01",
+    end_date = "2000-12-31",
+    clim_format_version = "v2"
+  )
+
+  expect_true(file.exists(res1$file[[1]]))
+  clim1 <- read.table(res1$file[[1]], header = FALSE, sep = "")
+  expect_equal(ncol(clim1), 14)
+  expect_true(all(clim1[, 1] == 0))
+  expect_true(all(clim1[, 14] == 0.6))
+
+  expect_true(file.exists(res2$file[[1]]))
+  clim2 <- read.table(res2$file[[1]], header = FALSE, sep = "")
+  expect_equal(ncol(clim2), 12)
+  expect_true(all(clim1[, 2:13] == clim2))
+})